requirements_gpu.txt
num2words
spacy
espeakng_loader
misaki[en]>=0.9.4
onnxruntime-gpu
soundfile
numpy
huggingface_hub
nvidia-cublas-cu12                                             
nvidia-cudnn-cu12                                                                                                                                                                    
nvidia-cuda-runtime-cu12                                       
nvidia-curand-cu12                                                                                                                                                                   
nvidia-cufft-cu12 
nvidia-cusolver-cu12                                                                                                                                                                 
nvidia-cusparse-cu12                                           
nvidia-nvjitlink-cu12
